requirements.txt
fastapi
uvicorn
python-dotenv
fastapi-mail
email-validator
pydantic
pydantic-settings
python-multipart
typing-extensions
starlette
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-jose[cryptography]
requests
python-jose==3.3.0